pyproject.toml
[project]
name = "linkedin-profile-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi[standard]>=0.116.1",
    "ir>=0.0.6",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.4",
    "logging>=0.4.9.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.42",
    "typing>=3.10.0.0",
    "uvicorn>=0.35.0",
]
